Fix ikey method to run on app insights not log analytics
